unify clearing deleted flow in app lifecycle and deltedProgramHandler
