+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/services/ApplicationLifecycleService.java
@@ -26,7 +26,6 @@ import co.cask.cdap.api.artifact.ArtifactSummary;
 import co.cask.cdap.api.artifact.ArtifactVersion;
 import co.cask.cdap.api.artifact.ArtifactVersionRange;
 import co.cask.cdap.api.flow.FlowSpecification;
-import co.cask.cdap.api.flow.FlowletConnection;
 import co.cask.cdap.api.metrics.MetricDeleteQuery;
 import co.cask.cdap.api.metrics.MetricStore;
 import co.cask.cdap.api.plugin.Plugin;
@@ -88,10 +87,8 @@ import com.google.common.base.Function;
 import com.google.common.base.Joiner;
 import com.google.common.base.Predicates;
 import com.google.common.base.Throwables;
-import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Maps;
-import com.google.common.collect.Multimap;
 import com.google.common.util.concurrent.AbstractIdleService;
 import com.google.gson.Gson;
 import com.google.inject.Inject;
@@ -107,7 +104,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import javax.annotation.Nullable;
 
@@ -705,8 +701,6 @@ public class ApplicationLifecycleService extends AbstractIdleService {
    * @throws Exception
    */
   private void deleteApp(final ApplicationId appId, ApplicationSpecification spec) throws Exception {
-    Id.Application idApplication = appId.toId();
-
     //Delete the schedules
     scheduler.deleteSchedules(appId);
 
@@ -716,34 +710,9 @@ public class ApplicationLifecycleService extends AbstractIdleService {
     deletePreferences(appId);
 
     // Delete all streams and queues state of each flow
-    // TODO: This should be unified with the DeletedProgramHandlerStage
     for (final FlowSpecification flowSpecification : spec.getFlows().values()) {
-      Id.Program flowProgramId = Id.Program.from(idApplication, ProgramType.FLOW, flowSpecification.getName());
-
-      // Collects stream name to all group ids consuming that stream
-      final Multimap<String, Long> streamGroups = HashMultimap.create();
-      for (FlowletConnection connection : flowSpecification.getConnections()) {
-        if (connection.getSourceType() == FlowletConnection.Type.STREAM) {
-          long groupId = FlowUtils.generateConsumerGroupId(flowProgramId, connection.getTargetName());
-          streamGroups.put(connection.getSourceName(), groupId);
-        }
-      }
-      // Remove all process states and group states for each stream
-      final String namespace = String.format("%s.%s", flowProgramId.getApplicationId(), flowProgramId.getId());
-
-      impersonator.doAs(appId, new Callable<Void>() {
-
-        // TODO: (CDAP-7326) since one worker or flow can only be ran by a single instance of APP, (also a single
-        // version), should delete flow for each version
-        @Override
-        public Void call() throws Exception {
-          for (Map.Entry<String, Collection<Long>> entry : streamGroups.asMap().entrySet()) {
-            streamConsumerFactory.dropAll(appId.getParent().stream(entry.getKey()), namespace, entry.getValue());
-          }
-          queueAdmin.dropAllForFlow(appId.flow(flowSpecification.getName()));
-          return null;
-        }
-      });
+      FlowUtils.clearDeletedFlow(impersonator, queueAdmin, streamConsumerFactory,
+                                 appId.flow(flowSpecification.getName()), (FlowSpecification) spec);
     }
 
     ApplicationSpecification appSpec = store.getApplication(appId);
